# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "This is vehicle spare parts dealers API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders/store": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Make Order",
                "description": "Create new order",
                "operationId": "orderStore",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "client_name": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "string"
                                    },
                                    "payment_mode": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/product/store": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product product",
                "description": "Product creation",
                "operationId": "productStore",
                "requestBody": {
                    "description": "Request body for creating product product",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "price": {
                                        "type": "decimal"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/product/update/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update Product",
                "description": "Update record of a Product ",
                "operationId": "productUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with example",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request body for updating product",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "price": {
                                        "type": "decimal"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/product/delete/{id}'": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product by Id",
                "description": "Delete product",
                "operationId": "productDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "product deleted successfully"
                    },
                    "404": {
                        "description": "product not found"
                    }
                }
            }
        },
        "/api/product/search/{name}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search product",
                "description": "Search product by name",
                "operationId": "productSearch",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of product to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "product not found"
                    }
                }
            }
        }
    }
}